==> analysis/04_Biodiversity.Rmd <==
---
title: "04_Biodiversity"
author: "Rena Fukuda"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/04_Biodiversity/")
```

# Setting the Environment
## Set seed
```{r set_seed}
set.seed(93285723)
```

## Load in libraries
```{r load_packages}
pacman::p_load(tidyverse, patchwork, iNEXT, phyloseq, install= FALSE)
```

## Load in data
```{r load_data}
load("data/02_Phyloseq/raw_preprocessed_physeq.RData")
row_preprocessed_physeq

# Intuition check
min(sample_sums(row_preprocessed_physeq))

# Load in meta data
metadata_df <-
  row_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

# Set colors
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")
```


# Goals

1. Calculate Hill Diversity
2. Evaluate rarefaction curves
3. Evaluate  diversity values
4.Make notes of specific samples & their seq depth.

# Calculate Diversity with iNEXT
```{r calc_div}
# prepare input data
iNEXT_input_df <- row_preprocessed_physeq %>% 
  otu_table() %>%
  data.frame 

dim(iNEXT_input_df)

# Run iNEXT to calculate Hill numbers
# Note: species in rows, samples in columns
# Set seed
#iNEXT_data <- iNEXT::iNEXT(iNEXT_input_df, q = c(0,1,2), datatype = "abundance") 

#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the diversity
```{r}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
```

# Plot diversity
```{r prep_for_plot}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)

# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
```

# Plot diversity
```{r plot_rarefaction}
# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(row_preprocessed_physeq)),
                     guide = FALSE) + 
  theme(legend.position = "none")
```

# Manually plot diversity
## Rarefaction curve
```{r iNEXT_manual}
iNEXT_manual_df <- iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = "Assemblage") %>%
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # join with metadata
  left_join(., metadata_df, by="names") %>%
  # Add colors
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
  
# Inspect
dim(iNEXT_manual_df)

# Plot rarefaction 
iNEXT_man_rf <- iNEXT_manual_df %>%
  # Filter out rows calculated by rarefaction
  dplyr::filter(Method == "Rarefaction") 

# Rarefaction by Order.q
ggplot(data = iNEXT_man_rf, aes(x = m, y= qD, color = station, group = names)) + 
  geom_line(stat = "identity") +
  facet_grid(~Order.q) +
  scale_color_manual(values = station_colors)

# Rarefaction by sample station and Order.q
ggplot(data = iNEXT_man_rf, aes(x = m, y= qD, color = station, group = names)) + 
  geom_line(stat = "identity") +
  facet_grid(cols = vars(station), rows = vars(Order.q), scales = "fixed") +
  scale_color_manual(values = station_colors)
```

# Environmental variable check
## Var check
```{r environmental_pairs, fig.width =6, fig.height=6}
head(metadata_df)

env_df <- 
  metadata_df %>%
  dplyr::select(names, water_tempC, DO_mgL)

head(env_df)

pairs(dplyr::select(env_df, -names), upper.panel = NULL)
```

## Diversity vs salinity
```{r div_vs_salinity, fig.width =6, fig.height=6}
iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = salinity_psu, y= qD)) +
  facet_wrap(~Order.q, scales = "free") +
  geom_point(aes(color = station)) +
  stat_smooth(method = "lm", formula = y ~ poly(x,2) ) + 
  labs(x = "Salinity (psu)", y="# of ASVs") +
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom") 
```

# Session information
```{r session_info}
devtools::session_info()
```
